refactor: centralize socket action constants in shared module for better maintainability
