+++ backend/Actions.js
@@ -1,17 +1,2 @@
-const Actions = {
-  JOIN: "join",
-  JOINED: "joined",
-  DISCONNECTED: "disconnected",
-  CODE_CHANGE: "code-change",
-  SYNC_CODE: "sync-code",
-  LEAVE: "leave",
-  SEND_CHAT: "send-chat",
-  CHAT: "chat",
-  CURSOR_CHANGE: "cursor-change",
-  LANGUAGE_CHANGE: "language-change",
-  KICK: "kick",
-  KICK_USER: "kick_user",
-  KICKED: "kicked",
-};
-
-module.exports = Actions;
+// Re-export from shared for backward compatibility if imported elsewhere
+module.exports = require("../shared/actions");
