+++ frontend/src/Action.js
@@ -1,14 +1,1 @@
-export const Actions = {
-  JOIN: "join",
-  JOINED: "joined",
-  DISCONNECTED: "disconnected",
-  CODE_CHANGE: "code-change",
-  SYNC_CODE: "sync-code",
-  LEAVE: "leave",
-  SEND_CHAT: "send-chat",
-  CHAT: "chat",
-  CURSOR_CHANGE: "cursor-change",
-  LANGUAGE_CHANGE: "language-change",
-  KICK_USER: "kick_user",
-  KICKED: "kicked",
-};
+export { Actions } from "@shared/actions";
