+++ frontend/tsconfig.json
@@ -14,6 +14,7 @@
     "skipLibCheck": true,
     "types": ["vite/client"]
   },
+  
   "include": ["src", "vite-env.d.ts"],
   "exclude": ["node_modules", "dist"]
 }
