+++ frontend/vite.config.js
@@ -1,10 +1,26 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
-import path from "path";
+import path, { dirname } from "path";
+import { fileURLToPath } from "url";
+
+// Provide __dirname in ESM
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
   define: {
     "process.env.IS_PREACT": JSON.stringify("true"),
+  },
+  resolve: {
+    alias: {
+      "@shared": path.resolve(__dirname, "../shared"),
+    },
+  },
+  server: {
+    fs: {
+      // Allow serving files from one level up to reach the shared folder
+      allow: [path.resolve(__dirname, ".."), path.resolve(__dirname, "../shared")],
+    },
   },
 });
